Extract code to make patch more readable
